Organize repository into frontend/backend structure and add best practice files
- Created frontend and backend directories and moved relevant files.
- Updated frontend import paths to reflect new structure.
- Added .gitignore for common exclusions.
- Added placeholder backend/requirements.txt with identified Python dependencies.
- Added placeholder frontend/package.json.
- Ensured documentation files are at the root.
- Attempted to run backend tests; 2 JWT-related tests are failing (status 422).
