+++ frontend/ProductDetailPage.jsx
